+++ ios/Classes/models/NotificationButton.swift
@@ -1,0 +1,13 @@
+//
+//  NotificationButton.swift
+//  flutter_foreground_task
+//
+//  Created by Woo Jin Hwang on 8/5/24.
+//
+
+import Foundation
+
+struct NotificationButton {
+  let id: String
+  let text: String
+}