+++ ios/Classes/service/BackgroundService.swift
@@ -9,7 +9,8 @@ import Flutter
 import Foundation
 import UserNotifications
 
-let NOTIFICATION_ID: String = "flutter_foreground_task/backgroundNotification"
+let NOTIFICATION_ID: String = "flutter_foreground_task/notification"
+let NOTIFICATION_CATEGORY_ID: String = "flutter_foreground_task/notification_category"
 let BG_ISOLATE_NAME: String = "flutter_foreground_task/backgroundIsolate"
 let BG_CHANNEL_NAME: String = "flutter_foreground_task/background"
 
@@ -17,6 +18,10 @@ let ACTION_TASK_START: String = "onStart"
 let ACTION_TASK_REPEAT_EVENT: String = "onRepeatEvent"
 let ACTION_TASK_DESTROY: String = "onDestroy"
 let ACTION_RECEIVE_DATA: String = "onReceiveData"
+
+let ACTION_NOTIFICATION_BUTTON_PRESSED = "onNotificationButtonPressed"
+let ACTION_NOTIFICATION_PRESSED = "onNotificationPressed"
+let ACTION_NOTIFICATION_DISMISSED = "onNotificationDismissed"
 
 @available(iOS 10.0, *)
 class BackgroundService: NSObject {
@@ -48,62 +53,61 @@ class BackgroundService: NSObject {
     }
   }
   
-  private let userNotificationCenter: UNUserNotificationCenter
-  private var isGrantedNotificationAuthorization: Bool = false
+  private let notificationCenter: UNUserNotificationCenter
+  private let notificationPermissionManager: NotificationPermissionManager
+  private var canReceiveNotificationResponse: Bool = false
 
-  private var showNotification: Bool = true
-  private var playSound: Bool = false
-  private var prevInterval: Int? = nil
-  private var currInterval: Int = 5000
-  private var prevIsOnceEvent: Bool? = nil
-  private var currIsOnceEvent: Bool = false
-  private var prevCallbackHandle: Int64? = nil
-  private var currCallbackHandle: Int64? = nil
-  private var notificationContentTitle: String = ""
-  private var notificationContentText: String = ""
+  private var notificationOptions: NotificationOptions
+  private var notificationContent: NotificationContent
+  private var prevBackgroundTaskOptions: BackgroundTaskOptions?
+  private var currBackgroundTaskOptions: BackgroundTaskOptions
+  private var prevBackgroundTaskData: BackgroundTaskData?
+  private var currBackgroundTaskData: BackgroundTaskData
   
   override init() {
-    userNotificationCenter = UNUserNotificationCenter.current()
+    notificationCenter = UNUserNotificationCenter.current()
+    notificationPermissionManager = NotificationPermissionManager()
+    notificationOptions = NotificationOptions.getData()
+    notificationContent = NotificationContent.getData()
+    currBackgroundTaskOptions = BackgroundTaskOptions.getData()
+    currBackgroundTaskData = BackgroundTaskData.getData()
     super.init()
-    // userNotificationCenter.delegate = self
   }
   
   func run(action: BackgroundServiceAction) {
-    let prefs = UserDefaults.standard
-
-    showNotification = prefs.bool(forKey: SHOW_NOTIFICATION)
-    playSound = prefs.bool(forKey: PLAY_SOUND)
-    prevInterval = currInterval
-    currInterval = prefs.integer(forKey: TASK_INTERVAL)
-    prevIsOnceEvent = currIsOnceEvent
-    currIsOnceEvent = prefs.bool(forKey: IS_ONCE_EVENT)
-    prevCallbackHandle = currCallbackHandle
-    currCallbackHandle = prefs.object(forKey: CALLBACK_HANDLE) as? Int64
-    notificationContentTitle = prefs.string(forKey: NOTIFICATION_CONTENT_TITLE) ?? notificationContentTitle
-    notificationContentText = prefs.string(forKey: NOTIFICATION_CONTENT_TEXT) ?? notificationContentText
+    notificationOptions = NotificationOptions.getData()
+    notificationContent = NotificationContent.getData()
+    prevBackgroundTaskOptions = currBackgroundTaskOptions
+    currBackgroundTaskOptions = BackgroundTaskOptions.getData()
+    prevBackgroundTaskData = currBackgroundTaskData
+    currBackgroundTaskData = BackgroundTaskData.getData()
     
     switch action {
       case .START:
-        requestNotificationAuthorization()
+        requestNotification()
         isRunningService = true
-        if let callbackHandle = currCallbackHandle {
+        if let callbackHandle = currBackgroundTaskData.callbackHandle {
           executeDartCallback(callbackHandle: callbackHandle)
         }
         break
       case .RESTART:
-        sendNotification()
+        requestNotification()
         isRunningService = true
-        if let callbackHandle = currCallbackHandle {
+        if let callbackHandle = currBackgroundTaskData.callbackHandle {
           executeDartCallback(callbackHandle: callbackHandle)
         }
         break
       case .UPDATE:
-        sendNotification()
+        requestNotification()
         isRunningService = true
-        if let callbackHandle = currCallbackHandle {
-          if prevCallbackHandle != callbackHandle {
+        if let callbackHandle = currBackgroundTaskData.callbackHandle {
+          if prevBackgroundTaskData?.callbackHandle != callbackHandle {
             executeDartCallback(callbackHandle: callbackHandle)
           } else {
+            let prevInterval = prevBackgroundTaskOptions?.interval
+            let currInterval = currBackgroundTaskOptions.interval
+            let prevIsOnceEvent = prevBackgroundTaskOptions?.isOnceEvent
+            let currIsOnceEvent = currBackgroundTaskOptions.isOnceEvent
             if prevInterval != currInterval || prevIsOnceEvent != currIsOnceEvent {
               startRepeatTask()
             }
@@ -116,47 +120,103 @@ class BackgroundService: NSObject {
           self.destroyFlutterEngine()
         }
         removeAllNotification()
-        isGrantedNotificationAuthorization = false
         isRunningService = false
         break
     }
   }
   
-  private func requestNotificationAuthorization() {
-    if showNotification {
-      let options = UNAuthorizationOptions(arrayLiteral: .alert, .sound)
-      userNotificationCenter.requestAuthorization(options: options) { success, error in
-        if let error = error {
-          print("Authorization error: \(error)")
-        } else {
-          if (success) {
-            self.isGrantedNotificationAuthorization = true
-            self.sendNotification()
-          } else {
-            print("Notification authorization denied.")
-          }
-        }
-      }
+  private func onMethodCall(call: FlutterMethodCall, result: @escaping FlutterResult) {
+    switch call.method {
+      case "startTask":
+        startBackgroundTask()
+      default:
+        result(FlutterMethodNotImplemented)
     }
   }
   
-  private func sendNotification() {
-    if isGrantedNotificationAuthorization && showNotification {
-      let notificationContent = UNMutableNotificationContent()
-      notificationContent.title = notificationContentTitle
-      notificationContent.body = notificationContentText
-      if playSound {
-        notificationContent.sound = UNNotificationSound.default
+  func userNotificationCenter(_ center: UNUserNotificationCenter, 
+                              _ response: UNNotificationResponse,
+                              _ completionHandler: @escaping () -> Void) {
+    // If it is not a notification requested by this plugin, the processing below is ignored.
+    if response.notification.request.identifier != NOTIFICATION_ID { return }
+    
+    // Prevents duplicate processing due to the `registrar.addApplicationDelegate`.
+    if !canReceiveNotificationResponse { return }
+    canReceiveNotificationResponse = false
+    
+    let actionId = response.actionIdentifier
+    if notificationContent.buttons.contains(where: { $0.id == actionId }) {
+      backgroundChannel?.invokeMethod(ACTION_NOTIFICATION_BUTTON_PRESSED, arguments: actionId)
+    } else if actionId == UNNotificationDefaultActionIdentifier {
+      backgroundChannel?.invokeMethod(ACTION_NOTIFICATION_PRESSED, arguments: nil)
+    } else if actionId == UNNotificationDismissActionIdentifier {
+      backgroundChannel?.invokeMethod(ACTION_NOTIFICATION_DISMISSED, arguments: nil)
+    }
+    
+    completionHandler()
+  }
+  
+  func userNotificationCenter(_ center: UNUserNotificationCenter, 
+                              _ notification: UNNotification,
+                              _ completionHandler: @escaping (UNNotificationPresentationOptions) -> Void) {
+    // If it is not a notification requested by this plugin, the processing below is ignored.
+    if notification.request.identifier != NOTIFICATION_ID { return }
+    
+    if notificationOptions.playSound {
+      completionHandler([.alert, .sound])
+    } else {
+      completionHandler([.alert])
+    }
+    
+    // Prevents duplicate processing due to the `registrar.addApplicationDelegate`.
+    canReceiveNotificationResponse = true
+  }
+  
+  private func setNotificationActions() {
+    var actions: [UNNotificationAction] = []
+    for button in notificationContent.buttons {
+      let action = UNNotificationAction(identifier: button.id, title: button.text)
+      actions.append(action)
+    }
+    
+    let category = UNNotificationCategory(
+      identifier: NOTIFICATION_CATEGORY_ID,
+      actions: actions,
+      intentIdentifiers: [],
+      options: .customDismissAction
+    )
+    
+    notificationCenter.setNotificationCategories([category])
+  }
+  
+  private func requestNotification() {
+    if !notificationOptions.showNotification {
+      return
+    }
+    
+    notificationPermissionManager.checkPermission { permission in
+      if permission == NotificationPermission.DENIED {
+        print("notification permission denied..")
+        return
       }
       
-      let request = UNNotificationRequest(identifier: NOTIFICATION_ID, content: notificationContent, trigger: nil)
-      userNotificationCenter.add(request, withCompletionHandler: nil)
+      let content = UNMutableNotificationContent()
+      content.title = self.notificationContent.title
+      content.body = self.notificationContent.text
+      content.categoryIdentifier = NOTIFICATION_CATEGORY_ID
+      if self.notificationOptions.playSound {
+        content.sound = .default
+      }
+      self.setNotificationActions()
+      
+      let request = UNNotificationRequest(identifier: NOTIFICATION_ID, content: content, trigger: nil)
+      self.notificationCenter.add(request, withCompletionHandler: nil)
     }
   }
   
   private func removeAllNotification() {
-    userNotificationCenter.removePendingNotificationRequests(withIdentifiers: [NOTIFICATION_ID])
-    userNotificationCenter.removeDeliveredNotifications(withIdentifiers: [NOTIFICATION_ID])
+    notificationCenter.removePendingNotificationRequests(withIdentifiers: [NOTIFICATION_ID])
+    notificationCenter.removeDeliveredNotifications(withIdentifiers: [NOTIFICATION_ID])
   }
   
   private func executeDartCallback(callbackHandle: Int64) {
@@ -234,14 +294,14 @@ class BackgroundService: NSObject {
   private func startRepeatTask() {
     stopRepeatTask()
     
-    if currIsOnceEvent {
+    if currBackgroundTaskOptions.isOnceEvent {
       backgroundChannel?.invokeMethod(ACTION_TASK_REPEAT_EVENT, arguments: nil) { _ in
         for listener in self.taskLifecycleListeners {
           listener.onTaskRepeatEvent()
         }
       }
     } else {
-      let timeInterval = TimeInterval(currInterval / 1000)
+      let timeInterval = TimeInterval(currBackgroundTaskOptions.interval / 1000)
       repeatTask = Timer.scheduledTimer(withTimeInterval: timeInterval, repeats: true) { _ in
         self.backgroundChannel?.invokeMethod(ACTION_TASK_REPEAT_EVENT, arguments: nil) { _ in
           for listener in self.taskLifecycleListeners {
@@ -255,36 +315,5 @@ class BackgroundService: NSObject {
   private func stopRepeatTask() {
     repeatTask?.invalidate()
     repeatTask = nil
-  }
-  
-  private func onMethodCall(call: FlutterMethodCall, result: @escaping FlutterResult) {
-    switch call.method {
-      case "startTask":
-        startBackgroundTask()
-      default:
-        result(FlutterMethodNotImplemented)
-    }
-  }
-  
-  func userNotificationCenter(_ center: UNUserNotificationCenter,
-                              didReceive response: UNNotificationResponse,
-                              withCompletionHandler completionHandler: @escaping () -> Void) {
-    // If it is not a notification requested by this plugin, the processing below is ignored.
-    if response.notification.request.identifier != NOTIFICATION_ID { return }
-    
-    completionHandler()
-  }
-  
-  func userNotificationCenter(_ center: UNUserNotificationCenter,
-                              willPresent notification: UNNotification,
-                              withCompletionHandler completionHandler: @escaping (UNNotificationPresentationOptions) -> Void) {
-    // If it is not a notification requested by this plugin, the processing below is ignored.
-    if notification.request.identifier != NOTIFICATION_ID { return }
-    
-    if playSound {
-      completionHandler([.alert, .sound])
-    } else {
-      completionHandler([.alert])
-    }
   }
 }